+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+
 - fixed output to be the same for tag command
+
 ## [0.11.1-dev.4] - 2017-11-20
 
 - Auto update .bit.map.json to semantic versions