+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fixed bit list command display for deprecated components
+- fixed bit show with compare flag to display dependencies
+
 ## [0.11.1-dev.4] - 2017-11-20
 
 - Auto update .bit.map.json to semantic versions