+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - fixed output to be the same for tag command
+- fixed bit list command display for deprecated components
+- fixed bit show with compare flag to display dependencies
 
 ## [0.11.1-dev.4] - 2017-11-20